devided and moved list.c into container/buff.c and container/list.c
